face/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
numpy==2.2.6
opencv-contrib-python==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
